requirements.txt
langgraph
langchain-community
langchain-anthropic
tavily-python
pandas
numpy
langsmith
langchain_groq
langchain
faiss-cpu
google-generativeai
langchain-google-genai
Flask
gunicorn
Jinja2
Werkzeug
MarkupSafe
flask_restful
requests
flask_cors